requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.23.0
opencv-python>=4.8.0
torch>=2.0.0
